refactor(AHWR-1761): move getLatestApplications into application-helper module #patch
