+++ app/lib/context-helper.test.js
@@ -16,7 +16,6 @@ import {
   ONLY_HERD,
   PI_HUNT_AND_DAIRY_FOLLOW_UP_RELEASE_DATE,
 } from "../constants/claim-constants.js";
-import { getApplicationsBySbi } from "../api-requests/application-api.js";
 import { getClaimsByApplicationReference } from "../api-requests/claim-api.js";
 import {
   clearEndemicsClaim,
@@ -27,6 +26,7 @@ import {
   setSessionData,
 } from "../session/index.js";
 import { createTempReference } from "./create-temp-ref.js";
+import { getLatestApplications } from "./application-helper.js";
 import { AHWR_SCHEME, POULTRY_SCHEME } from "ffc-ahwr-common-library";
 import { config } from "../config/index.js";
 import {
@@ -36,7 +36,7 @@ import {
   poultryClaimRoutes,
 } from "../constants/routes.js";
 
-jest.mock("../api-requests/application-api.js");
+jest.mock("./application-helper.js");
 jest.mock("../api-requests/claim-api.js");
 jest.mock("../session/index.js");
 jest.mock("./create-temp-ref.js");
@@ -188,353 +188,88 @@ describe("context-helper", () => {
   });
 
   describe("refreshApplications", () => {
-    const mockRequest = {};
-    it("returns new world application and relevant old world application", async () => {
-      const newWorld = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: "2025-05-01T00:00:00.000Z",
-      };
-      const oldWorld = {
-        type: "VV",
-        reference: "AHWR-1111-2222",
-        data: {
-          visitDate: "2025-04-15T00:00:00.000Z",
-        },
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([newWorld, oldWorld]);
-      const { latestEndemicsApplication, latestVetVisitApplication } = await refreshApplications(
-        "123456789",
-        mockRequest,
-      );
+    const mockRequest = { logger: {} };
 
-      expect(latestEndemicsApplication.reference).toBe("IAHW-1111-2222");
-      expect(latestVetVisitApplication.reference).toBe("AHWR-1111-2222");
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestEndemicsApplication",
-        newWorld,
-      );
+    beforeEach(() => {
+      jest.clearAllMocks();
+    });
+
+    it("calls getLatestApplications with the supplied sbi and request logger", async () => {
+      getLatestApplications.mockResolvedValueOnce({
+        latestEndemicsApplication: undefined,
+        latestPoultryApplication: undefined,
+        latestVetVisitApplication: undefined,
+      });
+
+      await refreshApplications("123456789", mockRequest);
+
+      expect(getLatestApplications).toHaveBeenCalledWith("123456789", mockRequest.logger);
+    });
+
+    it("writes the resolved applications to session and returns them", async () => {
+      const latestEndemicsApplication = { reference: "IAHW-1111-2222" };
+      const latestPoultryApplication = { reference: "POUL-3333-4444" };
+      const latestVetVisitApplication = { reference: "AHWR-5555-6666" };
+      getLatestApplications.mockResolvedValueOnce({
+        latestEndemicsApplication,
+        latestPoultryApplication,
+        latestVetVisitApplication,
+      });
+
+      const result = await refreshApplications("123456789", mockRequest);
+
       expect(setSessionData).toHaveBeenCalledWith(
         mockRequest,
         "endemicsClaim",
         "latestVetVisitApplication",
-        oldWorld,
+        latestVetVisitApplication,
       );
-    });
-
-    it("returns new world application and ignores not relevant old world application", async () => {
-      const newWorld = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: "2025-05-01T00:00:00.000Z",
-      };
-      const oldWorld = {
-        type: "VV",
-        reference: "AHWR-1111-2222",
-        data: {
-          visitDate: "2023-04-15T00:00:00.000Z",
-        },
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([newWorld, oldWorld]);
-      const { latestEndemicsApplication, latestVetVisitApplication } = await refreshApplications(
-        "123456789",
-        mockRequest,
-      );
-
-      expect(latestEndemicsApplication.reference).toBe("IAHW-1111-2222");
-      expect(latestVetVisitApplication).toBeUndefined();
       expect(setSessionData).toHaveBeenCalledWith(
         mockRequest,
         "endemicsClaim",
         "latestEndemicsApplication",
-        newWorld,
-      );
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestVetVisitApplication",
-        undefined,
-      );
-    });
-
-    it("returns new world application when no old world application", async () => {
-      const newWorld = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: "2025-05-01T00:00:00.000Z",
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([newWorld]);
-      const { latestEndemicsApplication, latestVetVisitApplication } = await refreshApplications(
-        "123456789",
-        mockRequest,
-      );
-
-      expect(latestEndemicsApplication.reference).toBe("IAHW-1111-2222");
-      expect(latestVetVisitApplication).toBeUndefined();
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestEndemicsApplication",
-        newWorld,
-      );
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestVetVisitApplication",
-        undefined,
-      );
-    });
-
-    it("returns nothing when old world application only", async () => {
-      const oldWorld = {
-        type: "VV",
-        reference: "AHWR-1111-2222",
-        data: {
-          visitDate: "2023-04-15T00:00:00.000Z",
-        },
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([oldWorld]);
-      const { latestEndemicsApplication, latestVetVisitApplication } = await refreshApplications(
-        "123456789",
-        mockRequest,
-      );
-
-      expect(latestVetVisitApplication).toBeUndefined();
-      expect(latestEndemicsApplication).toBeUndefined();
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestEndemicsApplication",
-        undefined,
-      );
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestVetVisitApplication",
-        undefined,
-      );
-    });
-
-    it("returns poultry application", async () => {
-      const poultry = {
-        type: "POUL",
-        reference: "POUL-1111-2222",
-        createdAt: "2025-05-01T00:00:00.000Z",
-      };
-      const newWorld = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: "2025-05-01T00:00:00.000Z",
-      };
-      const oldWorld = {
-        type: "VV",
-        reference: "AHWR-1111-2222",
-        data: {
-          visitDate: "2023-04-15T00:00:00.000Z",
-        },
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([newWorld, oldWorld, poultry]);
-      const { latestPoultryApplication, latestEndemicsApplication, latestVetVisitApplication } =
-        await refreshApplications("123456789", mockRequest);
-
-      expect(latestEndemicsApplication.reference).toBe("IAHW-1111-2222");
-      expect(latestPoultryApplication.reference).toBe("POUL-1111-2222");
-      expect(latestVetVisitApplication).toBeUndefined();
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestEndemicsApplication",
-        newWorld,
+        latestEndemicsApplication,
       );
       expect(setSessionData).toHaveBeenCalledWith(
         mockRequest,
         "poultryClaim",
         "latestPoultryApplication",
-        poultry,
+        latestPoultryApplication,
       );
+      expect(result).toEqual({
+        latestEndemicsApplication,
+        latestPoultryApplication,
+        latestVetVisitApplication,
+      });
+    });
+
+    it("writes undefined to session when no applications were resolved", async () => {
+      getLatestApplications.mockResolvedValueOnce({
+        latestEndemicsApplication: undefined,
+        latestPoultryApplication: undefined,
+        latestVetVisitApplication: undefined,
+      });
+
+      await refreshApplications("123456789", mockRequest);
+
       expect(setSessionData).toHaveBeenCalledWith(
         mockRequest,
         "endemicsClaim",
         "latestVetVisitApplication",
         undefined,
       );
-    });
-
-    it("excludes NOT_AGREED poultry applications when selecting latestPoultryApplication", async () => {
-      const rejectedPoultry = {
-        type: "POUL",
-        reference: "POUL-1111-2222",
-        createdAt: "2026-04-29T00:00:00.000Z",
-        status: "NOT_AGREED",
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([rejectedPoultry]);
-
-      const { latestPoultryApplication } = await refreshApplications("123456789", mockRequest);
-
-      expect(latestPoultryApplication).toBeUndefined();
+      expect(setSessionData).toHaveBeenCalledWith(
+        mockRequest,
+        "endemicsClaim",
+        "latestEndemicsApplication",
+        undefined,
+      );
       expect(setSessionData).toHaveBeenCalledWith(
         mockRequest,
         "poultryClaim",
         "latestPoultryApplication",
         undefined,
       );
-    });
-
-    it("prefers an AGREED poultry application over a more recent NOT_AGREED poultry application", async () => {
-      const rejectedPoultry = {
-        type: "POUL",
-        reference: "POUL-9999-9999",
-        createdAt: "2026-04-29T00:00:00.000Z",
-        status: "NOT_AGREED",
-      };
-      const agreedPoultry = {
-        type: "POUL",
-        reference: "POUL-1111-2222",
-        createdAt: "2026-01-15T00:00:00.000Z",
-        status: "AGREED",
-      };
-      // rejected first to mimic typical date-desc ordering returned by the backend
-      getApplicationsBySbi.mockResolvedValueOnce([rejectedPoultry, agreedPoultry]);
-
-      const { latestPoultryApplication } = await refreshApplications("123456789", mockRequest);
-
-      expect(latestPoultryApplication.reference).toBe("POUL-1111-2222");
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "poultryClaim",
-        "latestPoultryApplication",
-        agreedPoultry,
-      );
-    });
-
-    it("excludes NOT_AGREED endemics applications when selecting latestEndemicsApplication", async () => {
-      const rejectedEndemics = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: "2026-04-29T00:00:00.000Z",
-        status: "NOT_AGREED",
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([rejectedEndemics]);
-
-      const { latestEndemicsApplication } = await refreshApplications("123456789", mockRequest);
-
-      expect(latestEndemicsApplication).toBeUndefined();
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestEndemicsApplication",
-        undefined,
-      );
-    });
-
-    it("prefers an AGREED endemics application over a more recent NOT_AGREED endemics application", async () => {
-      const rejectedEndemics = {
-        type: "EE",
-        reference: "IAHW-9999-9999",
-        createdAt: "2026-04-29T00:00:00.000Z",
-        status: "NOT_AGREED",
-      };
-      const agreedEndemics = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: "2026-01-15T00:00:00.000Z",
-        status: "AGREED",
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([rejectedEndemics, agreedEndemics]);
-
-      const { latestEndemicsApplication } = await refreshApplications("123456789", mockRequest);
-
-      expect(latestEndemicsApplication.reference).toBe("IAHW-1111-2222");
-      expect(setSessionData).toHaveBeenCalledWith(
-        mockRequest,
-        "endemicsClaim",
-        "latestEndemicsApplication",
-        agreedEndemics,
-      );
-    });
-
-    it("filters poultry and endemics independently", async () => {
-      const rejectedPoultry = {
-        type: "POUL",
-        reference: "POUL-9999-9999",
-        createdAt: "2026-04-29T00:00:00.000Z",
-        status: "NOT_AGREED",
-      };
-      const agreedEndemics = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: "2026-01-15T00:00:00.000Z",
-        status: "AGREED",
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([rejectedPoultry, agreedEndemics]);
-
-      const { latestPoultryApplication, latestEndemicsApplication } = await refreshApplications(
-        "123456789",
-        mockRequest,
-      );
-
-      expect(latestPoultryApplication).toBeUndefined();
-      expect(latestEndemicsApplication.reference).toBe("IAHW-1111-2222");
-    });
-
-    it("returns AGREED applications unchanged when no NOT_AGREED is present", async () => {
-      const agreedEndemics = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: "2026-01-15T00:00:00.000Z",
-        status: "AGREED",
-      };
-      const agreedPoultry = {
-        type: "POUL",
-        reference: "POUL-1111-2222",
-        createdAt: "2026-01-15T00:00:00.000Z",
-        status: "AGREED",
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([agreedEndemics, agreedPoultry]);
-
-      const { latestEndemicsApplication, latestPoultryApplication } = await refreshApplications(
-        "123456789",
-        mockRequest,
-      );
-
-      expect(latestEndemicsApplication).toBe(agreedEndemics);
-      expect(latestPoultryApplication).toBe(agreedPoultry);
-    });
-
-    it("resolves latestVetVisitApplication against the AGREED endemics application's createdAt, ignoring a more recent NOT_AGREED one", async () => {
-      const oldWorldVisitDate = "2025-04-15T00:00:00.000Z";
-      const dateWithinTenMonthsOfOldWorldVisit = "2025-09-01T00:00:00.000Z";
-      const dateBeyondTenMonthsOfOldWorldVisit = "2026-05-01T00:00:00.000Z";
-      const rejectedEndemics = {
-        type: "EE",
-        reference: "IAHW-9999-9999",
-        createdAt: dateBeyondTenMonthsOfOldWorldVisit,
-        status: "NOT_AGREED",
-      };
-      const agreedEndemics = {
-        type: "EE",
-        reference: "IAHW-1111-2222",
-        createdAt: dateWithinTenMonthsOfOldWorldVisit,
-        status: "AGREED",
-      };
-      const oldWorld = {
-        type: "VV",
-        reference: "AHWR-1111-2222",
-        data: {
-          visitDate: oldWorldVisitDate,
-        },
-      };
-      getApplicationsBySbi.mockResolvedValueOnce([rejectedEndemics, agreedEndemics, oldWorld]);
-
-      const { latestEndemicsApplication, latestVetVisitApplication } = await refreshApplications(
-        "123456789",
-        mockRequest,
-      );
-
-      expect(latestEndemicsApplication.reference).toBe("IAHW-1111-2222");
-      expect(latestVetVisitApplication.reference).toBe("AHWR-1111-2222");
     });
   });
